+++ src/owmStructures/owmstruct.go
@@ -5,14 +5,14 @@ type OWMStruct struct {
 	Weather [1]Weather `json:"weather"`
 	Base string `json:"base"`
 	Main Main `json:"main"`
-	Visibility int `json:"visibility"`
+	Visibility int64 `json:"visibility"`
 	Wind Wind `json:"wind"`
 	Clouds Clouds `json:"clouds"`
-	Dt int `json:"dt"`
+	Dt int64 `json:"dt"`
 	Sys Sys `json:"sys"`
-	Timezone int `json:"timezone"`
-	Id int `json:"id"`
+	Timezone int64 `json:"timezone"`
+	Id int64 `json:"id"`
 	Name string `json:"name"`
-	Cod int `json:"cod"`
+	Cod int64 `json:"cod"`
 	Message string `json:"message"`
 }
