+++ src/owmStructures/uvstruct.go
@@ -4,6 +4,6 @@ type UVStruct struct {
 	Lat float64 `json:"lat"`
 	Lon float64 `json:"lon"`
 	Date_iso string `json:"date_iso"`
-	Date int `json:"date"`
+	Date int64 `json:"date"`
 	Value float64 `json:"value"`
 }
